third_party/blink/web_tests/http/tests/local/fileapi/file-last-modified-after-delete-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Test File.lastModifiedDate returns null if the information is not available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.dataTransfer contains a File object on drop.
PASS lastModifiedDate is not null
PASS lastModifiedDate is >= testStartTime
PASS (new Date()).getTime() is >= lastModifiedDate
PASS successfullyParsed is true

TEST COMPLETE
